requirements.txt
# Core web framework
streamlit==1.30.0

# LangChain ecosystem
langchain
langchain-community
langchain-core
langsmith
langchain-text-splitters==0.0.1

# Transformers & Embeddings
huggingface-hub==0.19.4
sentence-transformers==2.2.2
transformers==4.35.2

# FAISS (CPU version, prebuilt)
faiss-cpu==1.11.0

# OCR and PDF processing
pdf2image==1.16.3
pytesseract==0.3.10
pypdf==3.17.1
docx2txt==0.8

# Translation and NLP
deep-translator==1.11.4
nltk==3.8.1
markdownify==0.11.6

# Environment & utils
python-dotenv==1.0.0
tqdm==4.66.1

# Security & Auth
passlib==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1

# Data handling and ML
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
plotly==5.18.0

# Evaluation
rouge==1.0.1
flask==2.3.3
PyPDF2==3.0.1

numpy
spacy

sumy
bert-extractive-summarizer
deep-translator



docx2txt
bcrypt
python-jose
python-dotenv